rework to use labels instead of trying to use crystal waypoints
